Fix artifact selection bug (#53)
Artifact selection uses an outer loop over all assets, and an inner loop over possible valid names. This means that it will pick the first asset that matches any valid names, instead of the best match for the earliest name in the list.

The fix should just be to switch the loops!