Take into account architecture when selecting linux artifact (#59)
StyLua recently started producing linux-aarch64 builds. Foreman does not correctly take into account the architecture on linux, and selects the linux-aarch64 build when linux-x86_64 should be used. This causes GitHub Actions CI to fail.

We fix this by preferring a more explicit name in platform keywords for linux (depending on arch), and falling back to the generic one if nothing is found