LUAFDN-1603 - add support for auth format for artifactory use (#72)
Some roblox internal tools host builds on artifactory rather than via
GitHub/GitLab. This is the first step towards supporting artifactory as
a target for binary releases.

The auth format used will be standardized across several internal tools,
so the schema is included for posterity and used to validate the example
in the tests.