Add command to add Artifactory Auth (#91)
We can already read Artifactory tokens, but Foreman has no way of adding
the tokens them itself. This PR looks to add the command
`artifactory-auth`, that places the tokens in the expected format in the
expected location.

This PR also includes a minor version bump since we are adding behavior.